refactor: reorganize prompt components and rename AppErrorPrompt
- Move prompt components to common/prompt/ folder
- Rename AppErrorPrompt to ErrorPrompt
- Update imports and test IDs
