+++ x-pack/platform/plugins/shared/agent_builder/public/application/components/common/prompt/layout.tsx
@@ -53,12 +53,12 @@ export const PromptLayout: React.FC<PromptLayoutProps> = ({
 
   return (
     <EuiEmptyPrompt
-      data-test-subj="appErrorPrompt"
+      data-test-subj="ErrorPrompt"
       css={isEmbeddable ? undefined : defaultPromptStyles}
       hasShadow={!isEmbeddable}
       color={isEmbeddable ? 'transparent' : 'plain'}
       icon={iconContent}
-      title={<h2 data-test-subj="appErrorPromptTitle">{title}</h2>}
+      title={<h2 data-test-subj="ErrorPromptTitle">{title}</h2>}
       body={
         <EuiText color="subdued" textAlign="center">
           {subtitle}
